+++ CHANGELOG.md
@@ -9,11 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### 🚀 Added
 
-- TBD
+- Gracefully handle string annotation resolution errors in visitor
 
 ### ✨ Changed
 
-- TBD
+- Use regular dict for visitor dispatch cache. Weakref is not needed, since methods will always be alive as long as the class is alive.
+- Clarify visitor dispatch cache behavior & limitations in docstring
+- Expand visitor tests
 
 ### ⚠️ Breaking